+++ custom_model_runner/datarobot_drum/drum/perf_testing.py
@@ -30,6 +30,8 @@ from datarobot_drum.resource.drum_server_utils import DrumServerRun
 from datarobot_drum.resource.transform_helpers import (
     read_csv_payload,
     read_mtx_payload,
+    make_csv_payload,
+    make_mtx_payload,
 )
 
 
@@ -229,6 +231,23 @@ class CMRunTests:
 
         self._df_for_test = None
         self._test_cases_to_run = None
+
+    @staticmethod
+    def resolve_labels(target_type, options):
+        if target_type == TargetType.BINARY:
+            labels = [options.negative_class_label, options.positive_class_label]
+        elif target_type == TargetType.MULTICLASS:
+            labels = options.class_labels
+        else:
+            labels = None
+        return labels
+
+    @staticmethod
+    def load_transform_output(response, is_sparse, request_key):
+        if is_sparse:
+            return pd.DataFrame(read_mtx_payload(eval(response.text), request_key))
+        else:
+            return pd.DataFrame(read_csv_payload(eval(response.text), request_key))
 
     def _prepare_test_cases(self):
         print("Preparing test data...")
@@ -581,7 +600,7 @@ class CMRunTests:
         print("\n\nValidation checks results")
         print(tbl_report)
 
-    def check_prediction_side_effects(self):
+    def check_prediction_side_effects(self, target_temp_location=None):
         rtol = 2e-02
         atol = 1e-06
         input_extension = os.path.splitext(self.options.input)
@@ -591,58 +610,67 @@ class CMRunTests:
             df = pd.DataFrame(mmread(self.options.input).tocsr())
             samplesize = min(1000, max(int(len(df) * 0.1), 10))
             data_subset = df.sample(n=samplesize, random_state=42)
-            _, __tempfile_sample = mkstemp(suffix=".mtx")
-            sparse_mat = vstack(x[0] for x in data_subset.values)
-            mmwrite(__tempfile_sample, sparse_mat)
+            subset_payload = make_mtx_payload(data_subset)
         else:
             df = pd.read_csv(self.options.input)
             samplesize = min(1000, max(int(len(df) * 0.1), 10))
             data_subset = df.sample(n=samplesize, random_state=42)
-            _, __tempfile_sample = mkstemp(suffix=".csv")
-            data_subset.to_csv(__tempfile_sample, index=False)
+            subset_payload = make_csv_payload(data_subset)
 
-        if self.target_type == TargetType.BINARY:
-            labels = [self.options.negative_class_label, self.options.positive_class_label]
-        elif self.target_type == TargetType.MULTICLASS:
-            labels = self.options.class_labels
-        else:
-            labels = None
+        labels = self.resolve_labels(self.target_type, self.options)
 
         with DrumServerRun(
-            self.target_type.value,
-            labels,
-            self.options.code_dir,
+            self.target_type.value, labels, self.options.code_dir, verbose=False
         ) as run:
-            response_key = (
-                X_TRANSFORM_KEY
-                if self.target_type == TargetType.TRANSFORM
-                else RESPONSE_PREDICTIONS_KEY
-            )
-            endpoint = "/transform/" if self.target_type == TargetType.TRANSFORM else "/predict/"
 
-            response_full = requests.post(
-                run.url_server_address + endpoint, files={"X": open(self.options.input)}
-            )
+            endpoint = "/transform/" if self.target_type == TargetType.TRANSFORM else "/predict/"
+            payload = {"X": open(self.options.input)}
+            if self.target_type == TargetType.TRANSFORM:
+                if self.options.target:
+                    target_location = target_temp_location.name
+                    payload.update({"y": open(target_location)})
+                elif self.options.target_csv:
+                    target_location = self.options.target_csv
+                    payload.update({"y": open(target_location)})
+
+            response_full = requests.post(run.url_server_address + endpoint, files=payload)
+            if not response_full.ok:
+                raise DrumCommonException(
+                    "Failure in transform server: {}".format(response_full.text)
+                )
+
+            if is_sparse:
+                subset_payload = ("X.mtx", subset_payload)
 
             response_sample = requests.post(
-                run.url_server_address + endpoint, files={"X": open(__tempfile_sample)}
+                run.url_server_address + endpoint, files={"X": subset_payload}
             )
 
             if self.target_type == TargetType.TRANSFORM:
-                if is_sparse:
-                    preds_full = pd.DataFrame(read_mtx_payload(eval(response_full.text)))
-                    preds_sample = pd.DataFrame(read_mtx_payload(eval(response_sample.text)))
-                else:
-                    preds_full = read_csv_payload(eval(response_full.text))
-                    preds_sample = read_csv_payload(eval(response_sample.text))
+                preds_full = self.load_transform_output(
+                    response=response_full, is_sparse=is_sparse, request_key=X_TRANSFORM_KEY
+                )
+                preds_sample = self.load_transform_output(
+                    response=response_sample, is_sparse=is_sparse, request_key=X_TRANSFORM_KEY
+                )
             else:
-                preds_full = pd.DataFrame(json.loads(response_full.text)[response_key])
-                preds_sample = pd.DataFrame(json.loads(response_sample.text)[response_key])
+                preds_full = pd.DataFrame(json.loads(response_full.text)[RESPONSE_PREDICTIONS_KEY])
+                preds_sample = pd.DataFrame(
+                    json.loads(response_sample.text)[RESPONSE_PREDICTIONS_KEY]
+                )
 
             preds_full_subset = preds_full.iloc[data_subset.index]
 
             matches = np.isclose(preds_full_subset, preds_sample, rtol=rtol, atol=atol)
             if not np.all(matches):
+                if is_sparse:
+                    _, __tempfile_sample = mkstemp(suffix=".mtx")
+                    sparse_mat = vstack(x[0] for x in data_subset.values)
+                    mmwrite(__tempfile_sample, sparse_mat)
+                else:
+                    _, __tempfile_sample = mkstemp(suffix=".csv")
+                    data_subset.to_csv(__tempfile_sample, index=False)
+
                 message = """
                             Error: Your predictions were different when we tried to predict twice.
                             No randomness is allowed.
@@ -652,5 +680,3 @@ class CMRunTests:
                     preds_full_subset[~matches][:10], preds_sample[~matches][:10], __tempfile_sample
                 )
                 raise ValueError(message)
-            else:
-                os.remove(__tempfile_sample)