+++ custom_model_runner/datarobot_drum/resource/utils.py
@@ -55,7 +55,9 @@ def _create_custom_model_dir(
     return custom_model_dir
 
 
-def _exec_shell_cmd(cmd, err_msg, assert_if_fail=True, process_obj_holder=None, env=os.environ):
+def _exec_shell_cmd(
+    cmd, err_msg, assert_if_fail=True, process_obj_holder=None, env=os.environ, verbose=True
+):
     """
     Wrapper used by tests and validation to run shell command.
     Can assert that the command does not fail (usually used for tests)
@@ -79,10 +81,11 @@ def _exec_shell_cmd(cmd, err_msg, assert_if_fail=True, process_obj_holder=None, 
         process_obj_holder.out_stream = stdout
         process_obj_holder.err_stream = stderr
 
-    if len(stdout):
-        print("stdout: {}".format(stdout))
-    if len(stderr):
-        print("stderr: {}".format(stderr))
+    if verbose:
+        if len(stdout):
+            print("stdout: {}".format(stdout))
+        if len(stderr):
+            print("stderr: {}".format(stderr))
     if assert_if_fail:
         assert p.returncode == 0, err_msg